refactor(history): remove dependency to concrete message types
