+++ historicalData/getStartPeriod.ts
@@ -1,10 +1,12 @@
-import { HistoricalChartTypes } from './historicalDataRepository.js'
-import { type ChartType } from './getQueryStatement.js'
+import { HistoricalDataTimeSpans } from './HistoricalDataTimeSpans.js'
 
-export const getStartPeriod = (type: ChartType, startMS: number): string => {
-	const selectedType = HistoricalChartTypes[type]
+export const getStartPeriod = (
+	type: keyof typeof HistoricalDataTimeSpans,
+	startMS: number,
+): string => {
+	const selectedType = HistoricalDataTimeSpans[type]
 	if (selectedType === undefined)
-		throw new Error(`${type} is not a valid chart type`)
+		throw new Error(`${type} is not a valid time span`)
 
 	return `from_milliseconds(${startMS}) - ${selectedType.duration.replace(
 		/s$/,
